Fix Expression#readNumber not parsing till end
When reading a number which happens to be the last value, the original
implementation would fail to parse the next character and error.
However, by adding a check if the next element is the last value and cur
is nil already, we can tell that we have reached the end of the
expression and as such, can just parse whatever we got.

TODO to improve and streamline this logic