Add corner cases for operator-less exp
In expressions like x3 and 5 5 + 5, we need to parse the 3 and middle 5 as an infix token and fail. Without doing so, the program will just run and return x or 5 since the precedence of the variables/numbers/constants are -1 which causes an early return.

Focus on allow the program to fail with an error over failing silently and giving the inappropriate result